// frontend-bd/src/assets/interfaces/BreweryInterface.tsx
export interface BreweryInterface {
  adress: string;
  city: string;
  country: string;
  email: string;
  facebook: string;
  foundation_year: number;
  id: number;
  instagram: string;
  latitude: number;
  longitude: number;
  menu: string;
  name: string;
  postal_code: string;
  province: string;
  rate_beer: string;
  telephone: string;
  picture :string;
  untappd: string;
  website: string;
}